Move stub EventSub webhooks request/response members into dedicated static class.
